Implemented database storage interface
Currently supports adding new users (Should be done on join event, though it does check it the player exists already), and tracking their energy amounts. No built-in add/subtract, so pull the value, do your math, and store it again.
It's abstracted via the Persistence class, implement that to make a new handler.